[client] - checkpoint
