+++ integration/ts/tsconfig.vite.json
@@ -1,0 +1,4 @@
+{
+  "extends": "@synnaxlabs/tsconfig/vite.json",
+  "include": ["vite.config.ts"]
+}
